requirements.txt
prance>=23.6.21
jsonref>=1.1.0
pyyaml>=6.0
openai>=1.30.0
